run jobs in parallel for a full collection of saves and roms
This is a pure brute-force random walk. Once a basic navmesh is bootstrapped we can switch to a more intelligent search with that data to avoid walking on already known nodes. Will need some sort of trigger to avoid bouncing on a wall for the entire duration of a search.