Build a navmesh of the world by stumbling around
Also exploit what the explorers found on previous batches.

Not particularly efficient at covering the map. This iteration aims to find 1000 nodes in the navmesh. It runs 42 batches of 60 jobs each of which run for 5 minutes of emulated game time. This ends up taking my cluster about 3 hours to complete.